refactor: change interval type from int to double in t_ping_args
- Updated the interval field in the t_ping_args structure to use double for more precise timing.
- Adjusted related functions to handle double values for interval validation and assignment.
- Ensured compatibility with atof for parsing interval values from command line arguments.

+++ includes/ft_ping.h
@@ -27,7 +27,7 @@ typedef struct s_ping_args {
     char        *host;
     int         ttl;
     int         count;
-    int         interval;
+    double      interval;
     int         timeout;
     int         verbose;
 } t_ping_args;
@@ -38,7 +38,7 @@ typedef struct s_ping {
     int         sockfd;
     int         ttl;
     int         count;
-    int         interval;
+    double      interval;
     int         timeout;
     int         verbose;
     int         quiet;
